Feature likai 20180111



查看合并请求 !44

+++ SocialSecurity/SocService/SocService-Business/src/main/java/com/ciicsh/gto/afsupportcenter/socialsecurity/socservice/business/SsEmpTaskService.java
@@ -62,7 +62,7 @@ public interface SsEmpTaskService extends IService<SsEmpTask> {
      * @param comAccountId
      * @return
      */
-    Integer getSerial(Integer comAccountId);
+    Integer getSerial(Long comAccountId);
 
     /**
      * 批量查询任务单信息
